Pasarlo a un componente aparet y varios videos por idioma
